requirement.txt
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gliner==0.2.13
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
protobuf==5.28.3
pyparsing==3.2.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
shiboken6==6.8.0.2
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
